[pci] Remove some SetIrqMode asserts
These asserts existed to catch the error condition where a
process is changing IRQ modes and in the teardown it was
found that there were still registered handlers. This is a bad
state, however it's also the state the driver would see if a
driver devhost died. Since this code predates process teardown
and suspend paths, it makes sense to remove them now.

Change-Id: I7c3d43bc8b69b20b4434e2bf7366af0cb2d18453