[ledger] Correct EntrySize computation.
GetEntrySize and GetInlinedEntrySize, which are used to compute the size
of regular and inlined entries, respectively, used to not account for
some memory overhead (notably size of pointers for key/value). That has
lead to an incorrect computation of the message size in GetEntriesInline
when determening whether it still fits into the fidl message.

This change adds missing pieces to the GetEntrySize/GetInlinedEntrySize
computation and to the message size computation to ensure pagination works
correctly. It also adds a test for the corner case: writing and then
reading many small inlined entries.

LE-401 #done

Change-Id: I15f08236b751ea00c402b1598c72726025e3fe77