[mdns] poll ioctl_netc_get_if_info
The standalone mdns app uses ioctl_netc_get_if_info to enumerate
interfaces and isn't notified of link changes by netstack. This CL
updates the ioctl-based InterfaceMonitor so it checks once a minute
for link changes. This is needed so that mdns_standalone will
continue to operate through local ip address changes. This won't be
used in production.

Change-Id: I314379da6e58cbb489fb187f6ed9d7a0b2fe2aeb