[canary] Always return 32 bit magic values
On success magic() returns a 32 bit value, but on error it returns
UINT64_MAX. Fixing this to return a 32 bit always and UINT32_MAX on
error allows users to always rely on a 32 bit value for magic storage.
This prevents type errors in lower optimization levels in gcc and saves
space in the common case. static_asserts have been updated to reflect
that UINT32_MAX is no longer a valid magic value.

Change-Id: I9619141e8681ede8523f8d2fc2122aea20f8b9d9