[blobstore] Track reserved blocks and nodes
When a blob is first created, we allocate a node and blocks in memory
to write the blob contents to. The allocated resources should not be
persisted to disk until the client has finished writing all data to
the blob.

Previously there existed the possibility that upon completion of
another blob, allocated nodes/blocks of blobs still being written
would also be persisted to disk. Upon successful cleanup of the
blobs, these blocks would be released; however, in the event of a
disk failure, the disk would be left in an inconsistent state.

This change allows reserved nodes/blocks for blobs that are not yet
completed to be stored safely in memory, without potential compromise
of the integrity of on-disk structures.

ZX-1898 #done

Change-Id: Ic424951154e0b145b498aa777aa8f2ccc0a55d54