[kernel][vm] Make sys_process_(read|write)_memory() clamp to mapping end
Previously, process_(read|write)_memory() would read/write any amount
out of the VMO in which it was started. Instead, clamp the length to
the mapped region. This is still not a full solution, as an ideal API
would need to read from the correct VMO for each page in the region,
but it's slightly less unexpected behavior.

ZX-1631

Change-Id: I24d543fd867160479e8e59b2a4951b6a87774314