[utest][debugger-test] Fix flaky failure in checking port packet events
It is possible for the port packet to report both ZX_THREAD_RUNNING
and ZX_THREAD_TERMINATED if the zx_port_wait() call is slow, which
would cause a test failure.  This can be reproduced by adding a sleep
before read_exception() or by running Zircon with a single CPU.

Fix by accepting combinations of port packet events.

ZX-1463 #done

Change-Id: Ie800f0121b51e910ee5da77685a63e90b138d4e7