[wlan] Allow scanning while being associated
This is a bit of a hack. Embedding a scanState inside the associatedState and
proxying event loop events through it and forwarding them to the inner
scanning state.

I think scanning is stateful but shouldn't be
part of the connection state machine itself. We would probably
introduce an unassociated state and create a new scanning facility which
could use the same implementation we currently have, which different
state could use.

NET-424 #done

Change-Id: I69e5e05856ec4e3d6c0f432fcaae338d75db146c