[debugger] Clarify thread state API
Firms up the contract for zx_read|write_thread_state and makes the state more
explicit for each platform.

The general register struct has been made for the current platform which
simplifies much fo the code.

The sizes are converted to size_t for consistency with other object calls.

Forwarding #defines and typedefs are included for downstream users until they
can be updated.

This will likely need to be revised somewhat again when we add more thread
states, which I plan to do in a followup.

Change-Id: I57e8e9979d073c8efdfaeb2683f9f83494d19abf