[kernel][arm64] switch the kcounter increment to a relaxed atomic load/store for arm64
Relaxed atomic adds are about as fast as a regular one on arm64, and
this should reduce the amount of damage that a preemption at a bad time
would do as its incrementing a kcounter.

x86 doesn't really have a relaxed atomic, and they're fairly expensive,
but the regular increment is done with a single instruction, so the
risk there is fairly minimal (though non zero). The damage due to a bad
preempt should only result in a missed increment, however.

Change-Id: I3efaee16f1308b9f300e36f7b8295ae27c1ce24f