[wlan][common] Get valid CBW given wlan_channel_t
GetValidCbw() returns a valid CBW given an unsanitized wlan_channel_t
with respect to the provided primary 20Mhz channel.

A main use case is to convert a user level notion of "20, 40" MHz
numerabl value to a radio-speak CBW.

While this forced correction is critial in order to prevent miconfiguration
of underlying radio, and related crash event, it does not eliminate
a need to perform sanity check against the inputs.

NET-319

Change-Id: I9fb13c1b78844fa626ff7158900ff88b4e0e5ade