[fidl] generate C structure names in CamelCase
Use "Msg" and "Rsp" as postfixes instead of "_request" and "_response"
to both make these symbols a bit more compact but also match fidl's
canonical style for short names (which is CamelCase).  Also avoids
"request" which has another meaning in fidl.  Use "Evt" as postfix
for responses with no requests (Events).

Prefix method names with their Interface names.

Name the message header "hdr" instead of "header" for increased
terseness and decreased likelyhood of collision with parameter
names (which by fidl style are not abbreviated).

Part of naming ergonomics work for TO-704

Change-Id: I467337ec414d2f6c75b8976f8e77d4146e2803ac