[fidl] C bindings: update docs to match reality, and vice versa.
Link from docs to code and vice versa.

Return an error when decoding a struct that doesn't have the
specified number of handles.

Return an error when encoding a struct that contains a pointer
with value FIDL_HANDLE_PRESENT.

Fix up some bad formatting in otherwise-unchanged parts of the doc.

Add TODOs about closing handles when errors encountered.

TO-323

Change-Id: Iac398a996321f85cc513b475110056042f444ba6