[mdns] simplify address fixup logic (and fix bug)
Address fixup logic in InterfaceTransceiver was overly complex,
handling AAAA resources that are never generated. In addition,
fixups stopped for a given resource section when the first A
resource was found. Because we're not currently removing duplicate
resources from messages, this was causing invalid IpAddresses to
reach the message writer, which would dcheck.

Duplicate removal will happen in a future CL.

Change-Id: Icf0eff1928ab6f9c7f949c130e9ff22f2c1d3710