[audio][mixer] Fix a buffering bug.
Prior to this, outputs were not taking into account the fifo depth when
determining where the high water fill mark was relative to the harware read
pointer (but it was getting it correct for the low water mark).  This meant that
if an output had its low/high water marks set to [10, 20] mSec, and a FIFO depth
of 2 mSec (typical for USB), that it would actually be bouncing between [10, 18]
mSec instead of [10, 20] (and working a bit harder than it should have been
working) as a result.

This becomes particularly noticible if you turn these numbers down to [2, 4]
mSec with a FIFO depth of 2 mSec, and your scheduler timing is good enough that
you actually wake up again with 0 frames to produce and start to trip ASSERTs.

Fix the bug, and update the ASSERTs.  It is technically not an error to wake up
and have 0 frames to produce (although it should be impossible to have a
negative number of frames to produce).

Change-Id: If1a197ac44ccd0a3f5b592b90c6a8aec3af82322