[arm64][hypervisor] Correctly align vector tables
The EL2 vector tables were misaligned, so when the address of the
secondary table was loaded into VBAR_EL2 (which has its first 10 bits
marked RES0), we were never able to load the secondary vector table.
This also meant that the primary vector table was prone to be misaligned
too.

This fixes both vector tables by aligning them correctly, and switching
to adr_global to load their addresses.

Change-Id: I8840ea1db6bd59b6fd88c46a2e317cb980711a98