[vdso][buildid] Remove a generated header dependency from the vdso
This CL rearranges how the zx_system_get_version vdsocall obtains the
buildid of the system.

Prior to this, it gets it from a define in the generated
config-buildid.h header. This means that any time the git state of the
Zircon repository changes, the VDSO changes. Since a very large number
of things depend on the VDSO, this can cause spurious rebuilds just by
changing git state.

Now this constant is part of the const VDSO data. Changing git state
causes the kernel to rebuild, but no ABI surfaces change.

ZX-1686 #done

Change-Id: Ibdbcc75d8c720d2c473c9d01c0a9cb12a4c6a063