[syscalls] Increase ZX_CPRNG_DRAW_MAX_LEN to 4096
zx_prng_draw was originally conceived to be a low bandwidth way to
get high quality entropy to seed other PRNGs.  Accordingly,
ZX_CPRNG_DRAW_MAX_LEN was set to a rather low value of 256 bytes.
As this is now the interface to a securely seeded PRNG within the
kernel it makes sense to increase this limit (while keeping it from
getting so large as to affect the overall system).  A page at a
time (i.e. 4 kb) seems reasonable.

ZX-1536

Change-Id: Ie1d6819e469d3f15b2f73c62994cdaafa6c468b8