[kernel][x86] Whitelist exceptions that we expect userland can generate
Before this change, we have a "default:" case which treats all
remaining interrupts/exceptions as things that userland can generate
and which should be delivered to userland as Zircon exceptions.

Change that to be an explicit whitelist.  That makes it clearer what
we expect to handle here and reduces the risk that we handle some
interrupts/exceptions wrongly without noticing.

Change-Id: I710c8bbf4c748fbcfe4b4e8e85718d6882d1132b