[debugger] Allow threads to be single stepped.
This adds a new thread state parameter to put the thread in single-step
mode (or not). The architecture-specific exception handlers use this
flag to enable or disable single-stepping on the thread before returning
from the exception.

On x86 this sets FLAGS.TF ("trap flag"). On ARM64, this sets
MDSCR_EL1.SS which causes the process state to be set to "single step"
upon exception return.

An alternative would be to standardize on the flags to enable
single-stepping like on X86. On ARM this would be to define bit 21 or
the saved CPSR flags bit to be "single step" and then set MDSCR_EL1
accordingly before returning from the exception. On ARM32 this bit is
defined, but on ARM64 it is always undefined and you can only control
the single step state indirectly.

While this alternate approach would have enabled us to avoid the new
bool on the thread structure, using a reserved bit on the flags seemed
like a bad idea that might bite us in the future (even if it's very
unlikely). This approach is also much cleaner from an API perspective
since there is a clear step state that is toggleable, rather than
relying on CPU-specific behavior (ARM clears the SS bit after doing a
step, x86 doesn't, for example).

Change-Id: I1ef441aca6728d7837098290106fdab07e178c4c