[scripts] Add an `fx doctor` command
The goal of this script is to detect common issues with a Fuchsia
checkout. For example, on OS X the xcode command line tool
installation often lapses. Ensuring that `xcode select --install` is
run as part of a checkout or build is problematic: the step involves
manual input. Detecting that it needs to be run, however, is perfectly
mechanizable.

Change-Id: I58753494eeac45547707650ade0f4d4f4faf5dce