[style] Further fixing messaging on check-header-guards.py
The string matching in the fix routine can fix header guards that don't
match the file path. The check routine only matches against the expected
file path. The messaging in the case of a bad header guard would be a
little confusing since the bad header guard would be replaced with the
corrected header guard.

Change-Id: Iae7aecb20317e4de009b16ec4a0716bae4bd6ae6