[kernel][sched] Don't move pinned threads on cpu unplug
When handling cpu unplug, leave threads which are pinned to a specific
cpu on that cpu's run queue. If the cpu is only temporarily unavailable
(e.g. suspend), then pinned threads will start running again soon. If
the cpu is unplugged indefinitely, then they threads will not run
again, but that's a risk of depending on a specific resource.

ZX-1691 #done

Change-Id: I357ad10960f6a79df10584a24022de8b3018715c