[zxcrypt] Rewrite TestDevice a unit test helpers
There's some outstanding flake in the zxcrypt unit tests, but current
output from failed tests isn't sufficiently useful in isolating it.
This CL attempts to improve the reporting of unexpected errors by
converting the methods of zxcrypt::testing::TestDevice to be unit test
helpers as defined in unittest.h.

ZX-1659

Change-Id: Ie23b323e559099ce4b2527ff07a87aee669f4449