[debugger] Add simple debugger IPC stubs.
This small library will be shared by both the debug router and the debug
client for IPC purposes. The design of this system is described in
go/zxdb.

This includes some test code written as part of the prototype that uses
gtest. This is currently commented out because there currently isn't a good
unit test story for host code. These tests will need to be put into a different
framework when that's figured out, and then enabled.

Change-Id: I60d0c06a931c99fac8ce1bccec9f6a2b810ecf20