[libc] undef _POSIX_THREAD_PRIORITY_SCHEDULING
Posix allows for an implement to choose not to support thread scheduling
by ensuring that _POSIX_THREAD_PRIORITY_SCHEDULING is either not
defined, or defined to be -1.

 * http://pubs.opengroup.org/onlinepubs/9699919799/basedefs/unistd.h.html
 * https://linux.die.net/man/7/posixoptions

Undefine _POSIX_THREAD_PRIORITY_SCHEDULING and remove function
prototypes associated with those functions. These functions are not used
by any fuchsia based caller.

Change-Id: I865743b37f69d2ef5e76e041e9ef41b662286930