[zxcrypt] Refactor TestDevice
This CL refactors TestDevice's lifecycle, especially with regards to its
block fifo lifecycle management.

I am fairly certain the root cause of the test flake was the way I was
using |fdio_watch_directory| did not guarantee I wouldn't get the wrong
device when aggressively un/binding.  That method calls a provided
callback whenever files are added, removed, etc. to a directory.
Watching a directory in the /dev tree is the only way to be notified of
new devices being added or removed right now.  Notably, it immediately
invokes the callback with an 'add' event for each existing files.

Previously, since multiple devices could be added by binding a single
driver (i.e., the requested driver, auto-detected FVM drivers and a
mid-layer driver for each), the code watched the /dev/class/block
directory for an alias whose topological path ended in the right suffix.
This was racy since we might get an 'add' event for a device from a
previous test that hadn't quite unbound, and mtach the suffix.

The new approach recursively calls |fdio_watch_directory| for each
element in the topological path, thus guaranteeing we can only bind to
the new ramdisk created for each test.

ZX-1659 #comment Fixes the test flake

Change-Id: I3fbf571e5d81a528ecccb9067fe8742e1cf4a151