[modular] Also measure Operations in benchmarks.
Operations are already instrumented for tracing, so we can just as well
include measurement specifications for them in the benchmark tspec.

Next up will be to capture other interesting measures about operations,
for example:

* How full is the queue when an operation is scheduled?
* What other operation is an operation part of?
* How do operation queues relate?

Also, we really want to measure simply *all* Operations currently
known to be defined, and it's somewhat redundant to have to add them
explicitly and manually to the tspec. So that could change in the
future too. For now it's fine to seem them all together.

MI4-770 #comment Also measure Operations in benchmarks.
MI4-163 #comment Also measure Operations in benchmarks.

Change-Id: Ic8e69d9f3e58a0c692c972586e789016ef236249