[machina] Add volatile block wrapper.
Add a BlockDispatcher that retains written blocks in RAM. This is
intented as a quick and dirty way to use a block device image file that
is included with the guest package when writes are needed.

To enable just add the 'volatile' option to the block spec string. Ex:

  '/pkg/data/disk.img,ro,volatile'

The above spec will open disk.img read-only but will be exposed
read-write via a ramdisk overlay.

Change-Id: Ic452e5b98ff8a08d94e9d50c43104ac06d95165e