[manifest] Move manifest to directory and create a more zircon one
This changes the structure of Jiri manifest to resemble other system
layers. The minimal manifest includes only the repoisoty itself and
hooks while the zircon one also includes other external build
dependencies.

This is a reland of 24a056a4e0d367103644a477cf9dba6523f33a68. The
previous version placed imports in the project section which wasn't
caught by Jiri during validation but failed during build. This change
needs to be coordinated with the change to infra/config and followed
by a manual roll to Garnet to switch to the new manifest location.

Change-Id: Id641afbecc42977de2cc817b368073289aac910e