[arm64][hypervisor] Allocate a page for El2State
When we pass El2State from the host to EL2, we go from the host virtual
address to the 1:1 physical address mapping in EL2. This means that
EL2State can cross a page boundary unless we're careful, which would
cause us to write into the wrong memory if we're not careful.

Change-Id: I2ed2c0b0fe9c3981ac920f87e5307508cf3331ff