[debugger] Add more debugger pieces.
Add a stub "zxdb" application. This is compiled in both "host" and "target"
mode.

Adds a debug_ipc library that's compiled in both host and target mode. The
Zircon debug-ipc sources will be moved here in a subsequent step.

Adds a stream buffer and unit tests to the new debug_ipc directory. This will
be used by both ends of the communication channel between the client and agent
to buffer data from the socket so it can be parsed into messages. debug_router
tests are added that refer to the new test for this.

C++ standard library classes are now used in the exception handler in
preference for the FBL ones and lower-level Zircon threads.

The exception handler is hooked up to the stream buffer (this currently goes
nowhere). In a later phase this will be hooked up to the message parser.

Change-Id: I14fb6bea670c83b0347b9393300e3530f02bfba8