[debugger] Add a stub debug agent.
The main component is the skeleton of the exception handler for debugged
processes. This currently just prints messages rather than does any debugging.
It also listens on a socket. This socket will be filled out in a later pass
with the command channel from the debugger frontend. More information on the
design is in http://go/zxdb

The main program sets up the exception handler and runs "ps" attached to it to
test that the process exceptions are being caught.

Change-Id: I9711eeb6dafdb35901e8b69ae8a4337293520e1e