[utest][hypervisor] Break most of ulib dependency
This is so we can begin the process of removing ulib/hypervisor and
integrating everything directly into lib/machina. This will ease a lot
of the inter-layer dependency tangling we see.

For now, the test still has to have a dependency on ulib/hypervisor for
the decode tests, but a follow up change will move that across too.

Change-Id: Iea20400b4df02cecfa674bf79b8567a770891902