[bt][common] Add TaskDomain abstraction
TaskDomain is a mixim class for objects that dispatch tasks onto a
dedicated task runner (on which internal state may be accessed) and has
strict and potentially racy shutdown behavior.

A TaskDomain can be instantiated to run on any given TaskRunner. It may
also optionally spawn and manage a thread that exists for the duration
of its lifetime.

NET-440 #comment

Change-Id: I4540ee4f8c4206a1501b5c07170c057de1b5e49e