[guest] Fix scenic backend.
It appears that when we started creating our own ApplicationContext in
the guest, this did not play nice with how the ViewProviderApp exposes
the guest view.

Since we no longer use the initial thread to drive VCPU execution we
can just move the GuestView to use the main looper and remove the
thread created for the view.

Change-Id: I483a6d7a8eede99152e8b59ddedaa4c404bc5552