[bt][gatt] Make GATT live on its own thread
GATT is a core profile used by internal GAP procedures (such as service
discovery) and by higher level profiles. Since it is a complex
multi-client system, it doesn't make sense for it to be processed
serially with other GAP events.

For that reason, GATT is moved to a separate TaskDomain within the
bthost driver. Internals such as ATT bearer management are now
decoupled from the GAP system and exposed via a high-level
message-posting abstraction.

NET-440 #done

Change-Id: I35de9e41df8705b6dc42e5408a5e664be765397c