[debugger] Add debug_ipc.
The message reader/writer and protocol files are moved from Zircon and updated.

Serialization and deserialization functions for the existing protocol
definition is added. Since all current messages are one-directional, the
serialization and deserialization of each message type is in a different file
depending on whether the code should be at the client or the agent (target)
side.

Moved the existing stream buffer to the debug_ipc namespace for consistency.

Add tests for the new serialization code.

Change-Id: I6936283440b2ce6b1e0eee57142edd8a509cc84e