[media] MTWN-62 fix media_service crash on shutdown
media_service was frequently crashing on shutdown. There were two
reasons for this:

1) Media processing within a component such as a demux or decoder
   was not properly wound down as the parts of the component were
   deleted. As a result, posted tasks were running against deleted
   objects or objects that were no longer in the correct state for
   media processing.

2) Shared pointers are held by multiple threads, causing objects to
   be deleted on arbitrary threads. Objects, specifically ones
   dealing with FIDL, need to be deleted on specific threads. This
   was causing crashes.

The first problem is fixed in this CL by 'acquiring' each stage in
an in-proc graph and purging tasks queued for that stage before the
stage is destroyed.

The second problem is fixed by using a deleter in conjunction with
shared_ptr that deletes the shared object on the correct thread. This
deleter has been applied to the fidl producer and consumer.

Change-Id: I245726bd7ee60f59ff86aa12433ab0d2b50adbb1