[debugger] Hook up more debugger infrastructure.
Creates a "main dispatcher" class called DebugAgent. This gets data from the
exception handler and handles it accordingly.

Creates a separate target and host version of zxdb. On the target it will talk
to itself via an in-process socket. While this is slightly extra work (having
the remote one is a higher priority right now), the transport has not yet been
settled on so this unblocks some progress.

The host version currently sets up the exception handler thread, sleeps
briefly, then exits. This will be filled out with a simple console-based UI in
the future.

Cleans up the serialization and seserialization of arrays with some code in
protocol_helpers.*

Make the StreamBuffer bidirectional and hook up buffered writes to the
ExceptionHandler in the agent.

Makes the ExceptionHandler respond to quit messages rather than exit when
nothing is being debugged.

Add a Launch command to the protocol.

Change-Id: I8ca94640100ecb30803d3c777b1be09e74deaed9