[media] allow media players to read from fdio channel
This CL adds a new MediaService method CreateFilePlayer, which allows
the client to create a MediaPlayer using a zx::channel underlying an
fd rather than a SeekingReader.

Currently, playing a file on disk requires the client to create a
file reader (a SeekingReader implementation). File readers are vended
by media_service, which needs file access to create such a reader.
Later, we'll no longer support the creation of file readers from a
path, eliminating the sandboxing exception that currently gives
media_service access to file systems.

Change-Id: I813ca8e737e3515492624021dfef6badbf4cb772