[debugger] Allow launching from the debugger IPC.
Hooks up the debug_ipc message LaunchRequest to the DebugAgent. Moves the
launch code that was in the debug_agent main file for testing to its own file.

The ownership of how processes are watched changes a bit. Now the DebugAgent
maintains the list of all debugged processes. These now have their own class
which is currently very simple but will be where we hang things like breakpoint
information off of in the future. The exception handler just takes a handle to
watch rather than ownership of the zx::process object.

Removes the test code in debug_agent which manually attached to a process.
Currently this is manually exercised by running the target "zxdb" build
which sends the IPC command.

Since requests now come only from IPC and happen on the same thread that the
exception handler runs on, the threadsafe code in ExceptionHandler can be
removed.

Write stubs for getting process and thread listings. These are blocked on
exposing task-utils as public from Zircon.

Change-Id: Ic8cb1385db0c86fd9fdfb8fd71471a0841c61f1b