[fidl] Fix "recursion too deep" test cases and union field offsets.
1. Before this change, the recursion_message_type was not a valid FIDL message.
In particular the pointers in the maybe_recurse union were not being followed
because they weren't properly encoded as Pointer types.
Furthermore, pointers to Unions doesn't actually trigger the recursion code path
in the encoder, so I had to change them to pointers to Structs.

Previously the test was passing by accident because it was only asserting that
ZX_ERR_INVALID_ARGS was returned. Now we assert that we get the expected
error message, and we also check that things work as expected at
FIDL_RECURSION_DEPTH - 1 and then break at FIDL_RECURSION_DEPTH - 1.

2. Fix the incorrect assumption that union data will immediately follow
the tag. This bug was revealed in the process of doing 1.

3. Improve some error messages.

Change-Id: If027f11744aea10b52974c1ca5976b14da721fd8