[debugger] Export ulib/task-utils as public.
This will be used to fetch the process and thread trees as part of the
debugger.

This code is exported as a static library because it's fairly little code and
likely to be not commonly used enough to justify shipping as a separate .so.

Update readme to say the files are generated automatically (I tried to do it
manually for a few minutes since it looked like these were in the source tree).

Change-Id: Id8b13a21876bbf6384918be40ab404d0928be24b