[modular] Allow creation of empty stories.
This required changing the assumption that there will always be a "root"
Module (the one that was passed into CreateStory()). This functionality
was largely a convenience for clients with the exception of how Story
state was derived from the "root" Module.

Now, we keep track of the first Module in a running Story that changes
state, and use this Module to derive Story state. In practice, this will
always be the first Module that was added to the Story, since the first
state transition is triggered by adding the Module itself.

MI4-820 #done
MI4-820 #comment Allow creation of empty stories.

Change-Id: Ie8e25108617124ef579ca113bc2c692501ecd39a