[kernel][futex] Allocate FutexNode on the stack, not in ThreadDispatcher
This saves a bit of memory by reducing the size of ThreadDispatcher.

FutexNode only needs to be allocated while a thread is waiting on a
futex.  The rest of the time, the FutexNode isn't needed.

Note that FutexContext::FutexWait() already checks or enforces the
invariant that the FutexNode is not in use when FutexWait() returns:
 * If result == ZX_OK, it does DEBUG_ASSERT(!node.IsInQueue())
   (which ~FutexNode() also does).
 * Otherwise, it ensures that the futex is not in a list of waiters by
   calling UnqueueNodeLocked().

Change-Id: I9e3fe54d7bd118592542e1ba098fb6b156d752f8