[media] Add http URL support to media_service
This CL adds MediaService.CreateHttpPlayer and MediaPlayer.SetHttpUrl
to the media FIDL definitions. It also removes
MediaService.CreateNetworkReader.

SeekingReader is now used only for cases in which the client wants to
implement its own SeekingReader.

This change makes MediaPlayer a superset of NetMediaPlayer, so we no
longer need the latter.

Change-Id: If99e64e2117688b6f3327700bd70aa8b5718ece5