[fidl] Tell legacy FIDL generators what languages to generate for
The legacy fidl1 generators can generate code for C++, Dart, and
Python. The Dart legacy generator only works if the map data is
generated, which is controlled in our GN build by whether we're
targeting Rust or Dart. This means a build that does not target Rust or
Dart at all will still attempt to generate Dart code, which fails due
to this missing map.

This plumbs the generated language list through the fidl generator
frontend to the legacy generator via the confusingly named 'generators'
flag so we no longer attempt to generate Dart code at all in
configurations that don't have Dart.

Change-Id: If5bc7fd09055821bd5c177a74ae0c29fdbfe48f5