[debugger] Hook up process and threads listing.
Walks the object trees to retrieve the process/job tree and the threads for a
given process.

This duplicates some functionality in the zircon task-utils library. But that
library doesn't support enumerating only the threads of a given process and is
cumbersome and error-prone to translate that library's concept of iteration to
the tree structure the debug_ipc needs.

Unit tests are added for the new functionality.

Change-Id: If20c75216ed64e54f393c149fa1c77a4732351b8