[debugger] Add basic debug command input.
This is the infrastructure for basic command input, parsing and dispatch.
A handful of commands are "implemented" with empty implementations both as
examples and to test the infrastructure. The command syntax aims to be very
similar to lldb.

There are two layers:

1. A "client" layer that currently is mostly empty. This will contain the code
   for communicating with and managing the "object model" of the debugger.
   Future GUI frontends or scripting interfaces would interface with this
   layer.

2. A "console" layer which is the text input. This parses text commands, will
   call the client layer accordingly, and will print the results on the screen.

A zxdb test target is added to test this.

Change-Id: Ic3d9bfd47c1f6e65a7955f75389e6948d193ea61