[fidl] Extract the async loop impl choice out of fidl_bindings2_unittests
This changes the fidl_bindings2_unittests test files to be agnostic of
the implementation of the async dispatcher. Instead, they interact with
the loop through a fidl::test::AsyncLoopForTesting interface that can
be substituted out at compile time.  This way consumers of the C++ fidl
bindings can run the tests for the bindings on top of their own async
dispatcher, so long as they wire it up to the
fidl::test::AsyncLoopForTesting interface.

Change-Id: I4eab752957c7294e00ae4dd92483e4aa78e6870f