[media] add threadsafe callback joiner to bin/media/util
https://fuchsia.atlassian.net/browse/MTWN-91 is happening, because
stages can still be working when an adjacent stage is shut destroyed.
The graph should 'acquire' all stages in the graph before shutting
down any stages in the graph to prevent this.

Implementing this logic is much simpler using a callback joiner. A
callback joiner generates any number of callbacks, and calls another
callback when all those callbacks have run. This makes it easy to
determine when a bunch of parallel async tasks have completed.

Change-Id: Ie8a7da7134161ed087ab8d04d295ab37752928eb