[threads] Add a token to suspend threads.
Suspending a thread will will return a handle to a token that allows resuming
that thread exactly once. This ensures that different clients suspending and
resuming threads don't stomp on each other's state.

Currently this is implemented as a separate API call to facilitate landing.
In the future task_suspend/resume will use the new signature and the *_token
versions will be deleted.

Refcounting is added on ThreadDispatcher for suspend/resume calls. This
will affect the legacy suspend/resume calls.

ZX-1072

Change-Id: I9ecc63409e2e7dcea733f3ec58db50c1e8b0210b